prompt_flows/outlander_copilot/requirements.txt
promptflow>=1.15.0
promptflow-tools>=1.4.0
azure-search-documents>=11.4.0
openai>=1.0.0
python-dotenv>=1.0.0
